Reuse main app.py click setup and rename scenario-runner #61
This avoids code duplication.
